third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-create-basics.html

<!DOCTYPE html>
<title>Credential Manager: create() basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/webauth/authenticator.mojom.js"></script>
<script src="resources/authenticator-helpers.js"></script>
<script>

// Common mock values for the mockAuthenticator.
var challenge = new TextEncoder().encode("climb a mountain");

var public_key_rp = {
    id: "1098237235409872",
    name: "Acme"
};

var public_key_user = {
    id: new TextEncoder().encode("1098237235409872"),
    name: "avery.a.jones@example.com",
    displayName: "Avery A. Jones",
    icon: "https://pics.acme.com/00/p/aBjjjpqPb.png"
};

var public_key_parameters =  [{
    type: "public-key",
    alg: -7,
},];

var publicKey = {
    challenge,
    rp: public_key_rp,
    user: public_key_user,
    pubKeyCredParams: public_key_parameters,
    excludeCredentials: [],
};

var raw_id = new TextEncoder("utf-8").encode("rawId");
var id = btoa("rawId");
var client_data_json = new TextEncoder("utf-8").encode("clientDataJSON");
var attestation_object = new TextEncoder("utf-8").encode("attestationObject");

promise_test(_ => {
    mockAuthenticator.setRawId(raw_id);
    mockAuthenticator.setId(id);
    mockAuthenticator.setClientDataJson(client_data_json);
    mockAuthenticator.setAttestationObject(attestation_object);

    mockAuthenticator.setAuthenticatorStatus(
        webauth.mojom.AuthenticatorStatus.SUCCESS);

    return navigator.credentials.create({publicKey}).then(r => {
        assert_equals(r.id, id, "id");
        assert_true(r.rawId instanceof ArrayBuffer);
        assert_array_equals(new Uint8Array(r.rawId),
            raw_id, "rawId returned is the same");
        assert_true(r.response instanceof AuthenticatorAttestationResponse);
        assert_true(r.response.clientDataJSON instanceof ArrayBuffer);
        assert_array_equals(new Uint8Array(r.response.clientDataJSON),
            client_data_json, "clientDataJSON returned is the same");
        assert_true(r.response.attestationObject instanceof ArrayBuffer);
        assert_array_equals(new Uint8Array(r.response.attestationObject),
            attestation_object, "attestationObject returned is the same");
        assert_not_exists(r.response, 'authenticatorData');
        assert_not_exists(r.response, 'signature');
    });
}, "Verify that the mock returns the values we give it.");

promise_test(t => {
    return promise_rejects(t, "NotSupportedError",
        navigator.credentials.create());
}, "navigator.credentials.create() with no argument.");

promise_test(t => {
    mockAuthenticator.setAuthenticatorStatus(
      webauth.mojom.AuthenticatorStatus.CANCELLED);
    return promise_rejects(t, "NotAllowedError",
        navigator.credentials.create({ publicKey }));
}, "Verify that cancelled error returned by mock is properly handled.");

promise_test(t => {
    mockAuthenticator.setAuthenticatorStatus(
        webauth.mojom.AuthenticatorStatus.UNKNOWN_ERROR);
    return promise_rejects(t, "NotReadableError",
        navigator.credentials.create({ publicKey }));
}, "Verify that unknown error returned by mock is properly handled.");

promise_test(t => {
    mockAuthenticator.setAuthenticatorStatus(
        webauth.mojom.AuthenticatorStatus.NOT_ALLOWED_ERROR);
    return promise_rejects(t, "NotAllowedError",
        navigator.credentials.create({ publicKey }));
}, "Verify that not allowed error returned by mock is properly handled.");

promise_test(t => {
    mockAuthenticator.setAuthenticatorStatus(
        webauth.mojom.AuthenticatorStatus.NOT_SUPPORTED_ERROR);
    return promise_rejects(t, "NotSupportedError",
        navigator.credentials.create({ publicKey }));
}, "Verify that not supported error returned by mock is properly handled.");

promise_test(t => {
    mockAuthenticator.setAuthenticatorStatus(
        webauth.mojom.AuthenticatorStatus.SECURITY_ERROR);
    return promise_rejects(t, "SecurityError",
        navigator.credentials.create({ publicKey }));
}, "Verify that security error returned by mock is properly handled.");

promise_test(t => {
 mockAuthenticator.setAuthenticatorStatus(webauth.mojom.AuthenticatorStatus.NOT_IMPLEMENTED);
   return promise_rejects(t, "NotAllowedError",
      navigator.credentials.create({ publicKey }));
}, 'Verify that not implemented error returned by mock is properly handled.');

promise_test(t => {
    var publicKey = {
        // No challenge.
        rp: public_key_rp,
        user: public_key_user,
        pubKeyCredParams: public_key_parameters,
    };

    return promise_rejects(t, new TypeError(),
            navigator.credentials.create({publicKey}));
}, "navigator.credentials.create() with missing challenge");

promise_test(t => {
    var publicKey = {
        challenge,
        rp: public_key_rp,
        user: public_key_user,
        // No parameters.
    };

    return promise_rejects(t, new TypeError(),
            navigator.credentials.create({publicKey}));
}, "navigator.credentials.create() with missing parameters");

promise_test(t => {
    var custom_public_key = {
        challenge,
        // No rp.
        user: public_key_user,
        pubKeyCredParams: public_key_parameters,
    };

    return promise_rejects(t, new TypeError(),
            navigator.credentials.create({publicKey: custom_public_key}));
}, "navigator.credentials.create() with missing rp");

promise_test(t => {
    var custom_public_key = {
        challenge,
        rp: public_key_rp,
        // No user.
        pubKeyCredParams: public_key_parameters,
    };

    return promise_rejects(t, new TypeError(),
            navigator.credentials.create({publicKey: custom_public_key}));
}, "navigator.credentials.create() with missing user");

promise_test(_ => {
    mockAuthenticator.reset();
    mockAuthenticator.setRawId(raw_id);
    mockAuthenticator.setId(id);
    mockAuthenticator.setClientDataJson(client_data_json);
    mockAuthenticator.setAttestationObject(attestation_object);
    mockAuthenticator.setAuthenticatorStatus(
          webauth.mojom.AuthenticatorStatus.SUCCESS);

    var publicKey = {
        challenge,
        rp: { name: "Acme" },
        user: public_key_user,
        pubKeyCredParams: public_key_parameters,
    };

    return navigator.credentials.create({publicKey: publicKey}).then(r => {
        assert_equals(r.id, id, 'id');
        assert_true(r.rawId instanceof ArrayBuffer);
        assert_array_equals(new Uint8Array(r.rawId),
            raw_id, "rawId returned is the same");
        assert_true(r.response instanceof AuthenticatorAttestationResponse);
        assert_true(r.response.clientDataJSON instanceof ArrayBuffer);
        assert_array_equals(new Uint8Array(r.response.clientDataJSON),
            client_data_json, "clientDataJSON returned is the same");
        assert_true(r.response.attestationObject instanceof ArrayBuffer);
        assert_array_equals(new Uint8Array(r.response.attestationObject),
            attestation_object, "attestationObject returned is the same");
        assert_not_exists(r.response, 'authenticatorData');
        assert_not_exists(r.response, 'signature');
    });
}, "navigator.credentials.create() with missing rp.id");


promise_test(t => {
    var custom_public_key = {
        challenge,
        rp: { id: "1098237235409872" },
        user: public_key_user,
        pubKeyCredParams: public_key_parameters,
    };

    return promise_rejects(t, new TypeError(),
            navigator.credentials.create({publicKey: custom_public_key}));
}, "navigator.credentials.create() with missing rp.name");

promise_test(t => {
    var custom_public_key = {
        challenge,
        rp: public_key_rp,
        user: {
            name: "avery.a.jones@example.com",
            displayName: "Avery A. Jones",
            icon: "https://pics.acme.com/00/p/aBjjjpqPb.png"
        },
        pubKeyCredParams: public_key_parameters,
    };

    return promise_rejects(t, new TypeError(),
            navigator.credentials.create({publicKey: custom_public_key}));
}, "navigator.credentials.create() with missing user.id");

promise_test(t => {
    var custom_public_key = {
        challenge,
        rp: public_key_rp,
        user: {
            id: new TextEncoder().encode("1098237235409872"),
            displayName: "Avery A. Jones",
            icon: "https://pics.acme.com/00/p/aBjjjpqPb.png"
        },
        pubKeyCredParams: public_key_parameters,
        timeout: 60000,  // 1 minute
        excludeCredentials: [], // No already-registered credentials.
    };

    return promise_rejects(t, new TypeError(),
            navigator.credentials.create({publicKey: custom_public_key}));
}, "navigator.credentials.create() with missing user.name");

promise_test(t => {
    var custom_public_key = {
        challenge,
        rp: public_key_rp,
        user: {
            id: new TextEncoder().encode("1098237235409872"),
            name: "avery.a.jones@example.com",
            icon: "https://pics.acme.com/00/p/aBjjjpqPb.png"
        },
        pubKeyCredParams: public_key_parameters,
    };

    return promise_rejects(t, new TypeError(),
            navigator.credentials.create({publicKey: custom_public_key}));
}, "navigator.credentials.create() with missing user.displayName");

promise_test(_ => {
    mockAuthenticator.reset();
    mockAuthenticator.setRawId(raw_id);
    mockAuthenticator.setId(id);
    mockAuthenticator.setClientDataJson(client_data_json);
    mockAuthenticator.setAttestationObject(attestation_object);
    mockAuthenticator.setAuthenticatorStatus(
          webauth.mojom.AuthenticatorStatus.SUCCESS);

    var public_key = {
        challenge,
        rp: public_key_rp,
        user: {
            id: new TextEncoder().encode("1098237235409872"),
            name: "avery.a.jones@example.com",
            displayName: "Avery A. Jones",
        },
        pubKeyCredParams: public_key_parameters,
    };

    return navigator.credentials.create({publicKey: public_key}).then(r => {
        assert_equals(r.id, id, 'id');
        assert_true(r.rawId instanceof ArrayBuffer);
        assert_array_equals(new Uint8Array(r.rawId),
            raw_id, "rawId returned is the same");
        assert_true(r.response instanceof AuthenticatorAttestationResponse);
        assert_true(r.response.clientDataJSON instanceof ArrayBuffer);
        assert_array_equals(new Uint8Array(r.response.clientDataJSON),
            client_data_json, "clientDataJSON returned is the same");
        assert_true(r.response.attestationObject instanceof ArrayBuffer);
        assert_array_equals(new Uint8Array(r.response.attestationObject),
            attestation_object, "attestationObject returned is the same");
        assert_not_exists(r.response, 'authenticatorData');
        assert_not_exists(r.response, 'signature');
    });
}, "navigator.credentials.create() with missing user.icon");

</script>
